+++ src/big_luca/repository.rs
@@ -24,6 +24,11 @@ impl Repository {
 
     /// Insert a chat to database
     pub async fn insert_chat(&self, chat: ChatId) -> anyhow::Result<()> {
+        if self.is_subscribed(&chat).await? {
+            anyhow::bail!(
+                "Sei già iscritto a katanga. Se vuoi lucrare di più iscriviti ai miei corsi."
+            )
+        }
         Chat::new(chat)
             .insert(self.db.pool())
             .await
@@ -57,5 +62,14 @@ impl Repository {
                     })
                     .collect()
             })
+    }
+
+    /// Check whether `chat_id` is subscribed
+    async fn is_subscribed(&self, chat_id: &ChatId) -> anyhow::Result<bool> {
+        Ok(self
+            .get_subscribed_chats()
+            .await?
+            .iter()
+            .any(|x| x == chat_id))
     }
 }