+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
 - [Changelog](#changelog)
-  - [0.2.0](#012)
+  - [0.2.1](#021)
+  - [0.2.0](#020)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.2.1
+
+Released on 24/08/2022
+
+- New scheduled aphorism in the afternoon
+- Changed scheduled time
+- Check whether chat is already subscribed
 
 ## 0.2.0